add support for chalk/stderr (closes #301)
