+++ test/_supports-color.js
@@ -7,6 +7,12 @@ const DEFAULT = {
 		hasBasic: true,
 		has256: true,
 		has16m: true
+	},
+	stderr: {
+		level: 2,
+		hasBasic: true,
+		has256: true,
+		has16m: false
 	}
 };
